requirements.txt
django
python-dotenv
whitenoise
django-admin-volt==1.0.10
psycopg2-binary
six==1.16.0
JSON-log-formatter==0.5.2
django-hosts==5.2
djangorestframework==3.14.0
boto3==1.26.142
django-otp==1.2.1
celery==5.2.7
gunicorn==20.1.0
django-log-request-id==2.1.0
django-softdelete==0.10.5
django-ses==3.5.0
pytest